test(e2e): harness-dnd — new visible-mode harness, absorb dnd probe
- New scripts/harness-dnd.mjs (visible-mode launch CCSM_E2E_HIDDEN=0)
- Absorbs scripts/probe-e2e-dnd.mjs as case `dnd` (deleted per #72 no-skipped-e2e)
- Updates harness-ui.mjs and docs/e2e-runner.md notes to point at the new home

Why a separate harness instead of harness-ui: harness-ui shares one
hidden electron launch and #335 capability set has no per-case env
override. dnd-kit needs visible window for pointer hit-testing and
hover-expand timer. Visible mode at the harness level isolates the
~2s window pop instead of leaking into harness-ui.

Currently 1 case but stands as the future home for any other
visible-mode-only e2e cases.
